Candidates service
Set up the api client and base url. Implementation
of the candidate service to get a list of candidates
that have been registered.
The call gets a list of options that are the start index
and the count which is the amount of candidates we want
the API to return